# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_httpresp.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  backend_internal_httpresp.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/backend_internal_httpresp.ErrorBody'
    type: object
...
  internal_interviews.CreateInterviewRequest:
    properties:
      application_id:
        description: ApplicationID links this interview to a job application.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      mode:
        description: |-
          Mode determines who drives the conversation.
          Valid values: "assist", "autonomous".
        enum:
        - assist
        - autonomous
        example: autonomous
        type: string
    required:
    - application_id
    - mode
    type: object
  internal_interviews.InterviewEventRequest:
    properties:
      content:
        description: Content is the spoken text or transcription.
        example: I have 5 years of experience in Go
        type: string
      feedback:
        description: |-
          Feedback is the full evaluation payload. Only used when Type="feedback".
          Stored as raw JSON because the schema varies by evaluation model.
        type: object
      score:
        description: Score is the AI's evaluation (0.0–100.0). Only used when Type="score".
        example: 85.5
        type: number
      speaker:
        description: Speaker identifies who is talking. Use "candidate", "ai", or
          "system".
        enum:
        - candidate
        - ai
        - system
        example: candidate
        type: string
      status:
        description: |-
          Status is the new session status. Only used when Type="status".
          Must be a valid transition from the current status.
        enum:
        - pending
        - active
        - completed
        - cancelled
        example: active
        type: string
      timestamp:
        description: |-
          Timestamp is when this turn occurred. Use *time.Time so that
          omitting it produces nil (not "0001-01-01T00:00:00Z").
        example: "2026-06-19T10:30:00Z"
        type: string
      type:
        description: |-
          Type identifies the kind of event. Valid values:
          "transcript", "status", "score", "feedback".
        enum:
        - transcript
        - status
        - score
        - feedback
        example: transcript
        type: string
    required:
    - type
    type: object
  internal_interviews.InterviewListResponse:
    properties:
      interviews:
        items:
          $ref: '#/definitions/internal_interviews.InterviewResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 15
        type: integer
    type: object
  internal_interviews.InterviewResponse:
    properties:
      application_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      created_at:
        example: "2026-06-19T09:55:00Z"
        type: string
      ended_at:
        example: "2026-06-19T10:30:00Z"
        type: string
      external_session_id:
        example: livekit-session-123
        type: string
      feedback:
        type: object
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      mode:
        enum:
        - assist
        - autonomous
        example: autonomous
        type: string
      model:
        example: gpt-4o-realtime-preview
        type: string
      provider:
        example: openai_realtime
        type: string
      score:
        example: 85.5
        type: number
      started_at:
        example: "2026-06-19T10:00:00Z"
        type: string
      status:
        enum:
        - pending
        - active
        - completed
        - cancelled
        example: active
        type: string
      transcript:
        items:
          $ref: '#/definitions/internal_interviews.TranscriptEntry'
        type: array
      updated_at:
        example: "2026-06-19T10:30:00Z"
        type: string
    type: object
  internal_interviews.StartInterviewRequest:
    properties:
      model:
        description: |-
          Model is the specific model for the provider (e.g., "gpt-4o-realtime-preview").
          Empty string means "use config default".
        example: gpt-4o-realtime-preview
        type: string
      provider:
        description: |-
          Provider is the voice backend (e.g., "openai_realtime", "elevenlabs").
          Empty string means "use config default".
        example: openai_realtime
        type: string
    type: object
  internal_interviews.StopInterviewRequest:
    properties:
      reason:
        description: |-
          Reason is a free-text explanation for stopping (for audit trail).
          Optional — empty string is acceptable.
        example: user_cancelled
        type: string
    type: object
  internal_interviews.TranscriptEntry:
    properties:
      content:
        description: |-
          Content is the spoken text (for AI) or transcription (for candidate).
          May be empty for SpeakerSystem metadata entries.
        type: string
      id:
        description: |-
          ID uniquely identifies this transcript entry. Generated by Go code,
          not by PostgreSQL.
        type: string
      speaker:
        description: |-
          Speaker identifies who is talking. Use SpeakerCandidate, SpeakerAI,
          or SpeakerSystem constants — not raw strings.
        type: string
      timestamp:
        description: |-
          Timestamp is when this turn occurred (UTC). Set by the voice service
          at the moment audio is received or speech is synthesized.
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/ai-job-search-agent/issues
  description: Local-first AI-powered job search automation platform. Discovers, scores,
    and applies to jobs with tailored resumes and cover letters.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/ai-job-search-agent
  title: AI Job Search Agent API
  version: "1.0"
paths:
...
  /internal/interviews/{id}/events:
    post:
      consumes:
      - application/json
      description: Internal endpoint for voice service to report transcript, status,
        score, feedback
      parameters:
      - description: Interview UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Event payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviews.InterviewEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Event processed
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid interview ID or request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Interview not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "409":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      summary: Handle voice service events (internal)
      tags:
      - Interviews
  /interviews:
    get:
      consumes:
      - application/json
      description: Get paginated list of interview sessions with optional filters
      parameters:
      - description: Filter by application UUID
        in: query
        name: application_id
        type: string
      - description: Filter by status
        enum:
        - pending
        - active
        - completed
        - cancelled
        in: query
        name: status
        type: string
      - description: Filter by mode
        enum:
        - assist
        - autonomous
        in: query
        name: mode
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated interview list
          schema:
            $ref: '#/definitions/internal_interviews.InterviewListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List interviews
      tags:
      - Interviews
    post:
      consumes:
      - application/json
      description: Create a new interview session in pending status
      parameters:
      - description: Interview creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviews.CreateInterviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created interview session
          schema:
            $ref: '#/definitions/internal_interviews.InterviewResponse'
        "400":
          description: Invalid request body or mode
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create interview session
      tags:
      - Interviews
  /interviews/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed interview session including transcript
      parameters:
      - description: Interview UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Interview session with transcript
          schema:
            $ref: '#/definitions/internal_interviews.InterviewResponse'
        "400":
          description: Invalid interview ID
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Interview not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get interview by ID
      tags:
      - Interviews
  /interviews/{id}/start:
    post:
      consumes:
      - application/json
      description: Start the interview session and enqueue voice task for browser-agent
      parameters:
      - description: Interview UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Start options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviews.StartInterviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Started interview session
          schema:
            $ref: '#/definitions/internal_interviews.InterviewResponse'
        "400":
          description: Invalid interview ID or request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Interview not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "409":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start interview session
      tags:
      - Interviews
  /interviews/{id}/stop:
    post:
      consumes:
      - application/json
      description: Stop an active interview session gracefully
      parameters:
      - description: Interview UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Stop reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interviews.StopInterviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Interview stopped
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid interview ID or request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Interview not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "409":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stop interview session
      tags:
      - Interviews
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...